# clients/web/openapi.json
{
  "components": {
    "schemas": {
...
      "AssistantCapabilities": {
        "additionalProperties": false,
        "properties": {
          "active_runtime": {
            "$ref": "#/components/schemas/ModelRuntime"
          },
          "architecture": {
            "items": {
              "$ref": "#/components/schemas/AssistantArchitectureStage"
            },
            "title": "Architecture",
            "type": "array"
          },
          "capability_registry_version": {
            "title": "Capability Registry Version",
            "type": "string"
          },
          "data_scope": {
            "enum": [
              "synthetic",
              "production-shadow",
              "production-readonly"
            ],
            "title": "Data Scope",
            "type": "string"
          },
          "deterministic_controls": {
            "items": {
              "type": "string"
            },
            "title": "Deterministic Controls",
            "type": "array"
          },
          "implementation_status": {
            "const": "local_governed_prototype",
            "title": "Implementation Status",
            "type": "string"
          },
          "model_responsibilities": {
            "items": {
              "type": "string"
            },
            "title": "Model Responsibilities",
            "type": "array"
          },
          "operation_capabilities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Operation Capabilities",
            "type": "array"
          },
          "production_gaps": {
            "items": {
              "type": "string"
            },
            "title": "Production Gaps",
            "type": "array"
          },
          "prohibited_model_actions": {
            "items": {
              "type": "string"
            },
            "title": "Prohibited Model Actions",
            "type": "array"
          },
          "validated_milestones": {
            "items": {
              "$ref": "#/components/schemas/ValidationMilestone"
            },
            "title": "Validated Milestones",
            "type": "array"
          }
        },
        "required": [
          "implementation_status",
          "data_scope",
          "active_runtime",
          "architecture",
          "model_responsibilities",
          "deterministic_controls",
          "prohibited_model_actions",
          "validated_milestones",
          "production_gaps",
          "capability_registry_version"
        ],
        "title": "AssistantCapabilities",
        "type": "object"
      },
      "AssistantMessageRequest": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AssistantMessageRequest",
        "type": "object"
      },
...
      "AuditSummary": {
        "additionalProperties": false,
        "properties": {
          "audit_id": {
            "title": "Audit Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "data_source": {
            "title": "Data Source",
            "type": "string"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "query_fingerprint": {
            "title": "Query Fingerprint",
            "type": "string"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "audit_id",
          "created_at",
          "status",
          "operation",
          "row_count",
          "query_fingerprint",
          "data_source"
        ],
        "title": "AuditSummary",
        "type": "object"
      },
...
      "ForecastRequest": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "default": 1000,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "reference_date": {
            "format": "date",
            "title": "Reference Date",
            "type": "string"
          },
          "scheme_id": {
            "minimum": 0.0,
            "title": "Scheme Id",
            "type": "integer"
          },
          "target_date": {
            "format": "date",
            "title": "Target Date",
            "type": "string"
          }
        },
        "required": [
          "reference_date",
          "target_date",
          "scheme_id"
        ],
        "title": "ForecastRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HumanFeedbackRequest": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "adopted_response": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssistantResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "correction": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Correction",
            "type": "string"
          }
        },
        "required": [
          "correction",
          "accepted"
        ],
        "title": "HumanFeedbackRequest",
        "type": "object"
      },
...
      "SessionRecord": {
        "additionalProperties": false,
        "properties": {
          "access_scope_hash": {
            "title": "Access Scope Hash",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "messages": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Messages",
            "type": "array"
          },
          "owner_subject_id": {
            "title": "Owner Subject Id",
            "type": "string"
          },
          "owner_tenant_or_department": {
            "title": "Owner Tenant Or Department",
            "type": "string"
          },
          "policy_snapshot_id": {
            "title": "Policy Snapshot Id",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "created_at",
          "owner_subject_id",
          "owner_tenant_or_department",
          "access_scope_hash",
          "policy_snapshot_id"
        ],
        "title": "SessionRecord",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Governed passenger-flow query API with a synthetic baseline",
    "title": "Metro Passenger Flow API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assistant/capabilities": {
      "get": {
        "operationId": "assistant_capabilities_api_v1_assistant_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantCapabilities"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assistant Capabilities",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/runs/{run_id}": {
      "get": {
        "operationId": "assistant_run_api_v1_assistant_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^run-[0-9a-f]{32}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assistant Run",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/runs/{run_id}/events": {
      "get": {
        "operationId": "assistant_events_api_v1_assistant_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^run-[0-9a-f]{32}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Assistant Events Api V1 Assistant Runs  Run Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assistant Events",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/runs/{run_id}/feedback": {
      "post": {
        "operationId": "assistant_feedback_api_v1_assistant_runs__run_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^run-[0-9a-f]{32}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assistant Feedback",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/sessions": {
      "post": {
        "operationId": "create_assistant_session_api_v1_assistant_sessions_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRecord"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Assistant Session",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/sessions/{session_id}/messages": {
      "post": {
        "operationId": "assistant_message_api_v1_assistant_sessions__session_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "pattern": "^session-[0-9a-f]{32}$",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assistant Message",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/audits/{audit_id}": {
      "get": {
        "operationId": "audit_api_v1_audits__audit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "audit_id",
            "required": true,
            "schema": {
              "pattern": "^(?:query|forecast)-[0-9a-f]{32}$",
              "title": "Audit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit",
        "tags": [
          "audits"
        ]
      }
    },
    "/api/v1/catalog": {
      "get": {
        "operationId": "catalog_api_v1_catalog_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Catalog",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/forecasts/designated-day": {
      "post": {
        "operationId": "forecast_api_v1_forecasts_designated_day_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Forecast",
        "tags": [
          "forecasts"
        ]
      }
    },
    "/api/v1/governance/status": {
      "get": {
        "operationId": "governance_status_api_v1_governance_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernanceStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Governance Status",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/queries": {
      "post": {
        "operationId": "query_api_v1_queries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Query",
        "tags": [
          "queries"
        ]
      }
    },
...
  }
}